requirements.txt
django>=1.9
Pillow

# For testing
pytest
pytest-flakes
pytest-cov
pytest-pep8
pytest-django
pyflakes
coverage

# frontend
django-bootstrap3
django-bootstrap3-datetimepicker-2
fuzzywuzzy